Cleaned up the workspace / files that I have edited.
I thought the issue with the exploding fetch model was the joint types. I thought the issue was related to damping not being registered by continuous joints and tried testing regular damping with revolute joints, however that has not fixed it. Setting the damping to 0 fixes it, but I think the point is supposed to add a level of friction to the joint movements to avoid "rag doll" behavior.